Decouple memory budget (buffer_bytes) from tensor_generator
